refactor: extract hmpb-ui library
Pulls it into a monorepo package visible to the React application. Adds tests to get reasonable coverage of the components extracted.
