+++ apps/client/src/config/globals.ts
@@ -1,6 +1,5 @@
 import { Dictionary } from '@votingworks/ballot-encoder'
 
-export const CHECK_ICON = '✓'
 export enum YES_NO_VOTES {
   no = 'No',
   yes = 'Yes',
