+++ apps/client/src/screens/OvervoteCombinationReportScreen.tsx
@@ -2,6 +2,7 @@ import React, { useContext } from 'react'
 import styled from 'styled-components'
 import pluralize from 'pluralize'
 
+import { Prose, Text, HorizontalRule } from '@votingworks/hmpb-ui'
 import NavigationScreen from '../components/NavigationScreen'
 import PrintButton from '../components/PrintButton'
 import LinkButton from '../components/LinkButton'
@@ -13,13 +14,10 @@ import {
   ContestTallyMeta,
 } from '../lib/votecounting'
 import Table, { TD } from '../components/Table'
-import Prose from '../components/Prose'
-import Text from '../components/Text'
 import {
   localeWeedkayAndDate,
   localeLongDateAndTime,
 } from '../utils/IntlDateTimeFormats'
-import HorizontalRule from '../components/HorizontalRule'
 
 const ContestMeta = styled.div`
   float: right;
