+++ apps/client/src/screens/TestDeckScreen.tsx
@@ -6,13 +6,13 @@ import {
   Precinct,
   VotesDict,
 } from '@votingworks/ballot-encoder'
+import { Prose } from '@votingworks/hmpb-ui'
 import routerPaths from '../routerPaths'
 
 import AppContext from '../contexts/AppContext'
 
 import PrintButton from '../components/PrintButton'
 import ButtonList from '../components/ButtonList'
-import Prose from '../components/Prose'
 import ContestTally from '../components/ContestTally'
 
 import {
