+++ apps/client/src/screens/UnconfiguredScreen.tsx
@@ -3,6 +3,7 @@ import { useHistory, useLocation } from 'react-router-dom'
 import styled from 'styled-components'
 import { parseElection } from '@votingworks/ballot-encoder'
 
+import { Prose, HorizontalRule } from '@votingworks/hmpb-ui'
 import ConverterClient, { VxFile } from '../lib/ConverterClient'
 import readFileAsync from '../lib/readFileAsync'
 
@@ -15,8 +16,6 @@ import AppContext from '../contexts/AppContext'
 import Button from '../components/Button'
 import routerPaths from '../routerPaths'
 import FileInputButton from '../components/FileInputButton'
-import HorizontalRule from '../components/HorizontalRule'
-import Prose from '../components/Prose'
 import Loading from '../components/Loading'
 import NavigationScreen from '../components/NavigationScreen'
 import Modal from '../components/Modal'
