+++ apps/client/tsconfig.json
@@ -1,9 +1,5 @@
 {
   "extends": "../../tsconfig.json",
-  "include": [
-    "src",
-    "test",
-    "types"
-  ],
+  "include": ["src", "test", "types", "../../libs/hmpb-ui/src/global.d.ts"],
   "compilerOptions": {}
 }
