+++ libs/hmpb-ui/.eslintrc.js
@@ -1,0 +1,142 @@
+const jsExtensions = ['.js', '.jsx']
+const tsExtensions = ['.ts', '.tsx']
+const allExtensions = jsExtensions.concat(tsExtensions)
+
+module.exports = {
+  env: {
+    browser: true,
+    node: true,
+    es6: true,
+    'jest/globals': true,
+    'cypress/globals': true,
+  },
+  parser: '@typescript-eslint/parser', // Specifies the ESLint parser
+  extends: [
+    'airbnb',
+    'plugin:import/errors',
+    'plugin:import/warnings',
+    'plugin:import/typescript',
+    'eslint:recommended',
+    'plugin:@typescript-eslint/recommended', // Uses the recommended rules from @typescript-eslint/eslint-plugin
+    'prettier/@typescript-eslint', // Uses eslint-config-prettier to disable ESLint rules from @typescript-eslint/eslint-plugin that would conflict with prettier
+    'plugin:prettier/recommended', // Enables eslint-plugin-prettier and displays prettier errors as ESLint errors. Make sure this is always the last configuration in the extends array.
+    'plugin:react/recommended', // Uses the recommended rules from @eslint-plugin-react
+    'plugin:jsx-a11y/recommended',
+    'prettier/react', // Overrides some of the rules in 'airbnb' to have more relaxed formatting in react.
+    'plugin:cypress/recommended',
+  ],
+  globals: {
+    Atomics: 'readonly',
+    SharedArrayBuffer: 'readonly',
+    globalThis: true,
+  },
+  parserOptions: {
+    ecmaFeatures: {
+      jsx: true,
+    },
+    ecmaVersion: 2018,
+    project: './tsconfig.json',
+    sourceType: 'module',
+  },
+  plugins: ['@typescript-eslint', 'jest', 'react', 'cypress', 'jsx-a11y'],
+  settings: {
+    react: {
+      version: 'detect', // Tells eslint-plugin-react to automatically detect the version of React to use
+    },
+    'import/extensions': allExtensions,
+    'import/parsers': {
+      '@typescript-eslint/parser': tsExtensions,
+    },
+    'import/resolver': {
+      node: {
+        extensions: allExtensions,
+      },
+      typescript: {
+        alwaysTryTypes: true,
+      },
+    },
+  },
+  rules: {
+    'jsx-a11y/label-has-associated-control': [
+      2,
+      {
+        assert: 'htmlFor',
+      },
+    ],
+    camelcase: 'error',
+    'import/no-extraneous-dependencies': [
+      'error',
+      {
+        devDependencies: true,
+      },
+    ],
+    'import/prefer-default-export': 'off',
+    'no-unused-vars': 'off', // base rule must be disabled as it can report incorrect errors: https://github.com/typescript-eslint/typescript-eslint/blob/master/packages/eslint-plugin/docs/rules/no-unused-vars.md#options
+    '@typescript-eslint/no-unused-vars': [
+      'error',
+      {
+        vars: 'all',
+      },
+    ],
+    quotes: ['error', 'single', { avoidEscape: true }],
+    'react/destructuring-assignment': 'off',
+    'react/jsx-boolean-value': [2, 'never'],
+    //
+    // ADDED RULES for election-manager
+    //
+    'react/require-default-props': 'off',
+    'react/prop-types': 'off',
+    //
+    // FIXME: address these new rules, part of eslint v6, required by this version of React.
+    'class-methods-use-this': 'off',
+    'no-shadow': 'off',
+    'no-param-reassign': 'off',
+    'no-nested-ternary': 'off',
+    'no-return-await': 'off',
+    'no-underscore-dangle': 'off',
+    'no-unused-expressions': 'off',
+    'consistent-return': 'off',
+    'no-multi-assign': 'off',
+    'guard-for-in': 'off',
+    'default-case': 'off',
+    'no-plusplus': 'off',
+    'no-await-in-loop': 'off',
+    'no-useless-constructor': 'off',
+    'no-return-assign': 'off',
+    //
+    //
+    'react/jsx-filename-extension': [
+      1,
+      {
+        extensions: ['.js', '.jsx', '.ts', '.tsx'],
+      },
+    ],
+    strict: 'off',
+    '@typescript-eslint/explicit-function-return-type': 'off', // Want to use it, but it requires return types for all built-in React lifecycle methods.
+    '@typescript-eslint/no-non-null-assertion': 'off',
+    '@typescript-eslint/no-use-before-define': ['error', { functions: false }],
+    'import/extensions': 'off',
+    'no-restricted-syntax': [
+      'error',
+      {
+        selector: 'Literal[value=null]',
+        message:
+          '`null` is generally not what you want outside of React classes, use `undefined` instead',
+      },
+    ],
+  },
+  overrides: [
+    {
+      files: 'cypress/**/*',
+      parserOptions: {
+        project: './cypress/tsconfig.json',
+      },
+    },
+    {
+      files: '**/*.{jsx,tsx}',
+      rules: {
+        'no-restricted-syntax': 'off',
+      },
+    },
+  ],
+}
