+++ libs/hmpb-ui/package.json
@@ -1,0 +1,38 @@
+{
+  "name": "@votingworks/hmpb-ui",
+  "version": "0.0.0-development",
+  "main": "index.js",
+  "license": "GPL-3.0",
+  "scripts": {
+    "lint": "eslint src/**/*.{ts,tsx}",
+    "test": "jest",
+    "test:watch": "jest --watch",
+    "test:coverage": "jest --coverage"
+  },
+  "peerDependencies": {
+    "dompurify": "^2.0.15",
+    "moment": "^2.27.0",
+    "pagedjs": "^0.1.42",
+    "react-dom": "^16.13.1",
+    "styled-components": "^5.2.0"
+  },
+  "devDependencies": {
+    "@testing-library/react": "^11.0.2",
+    "@types/faker": "^5.1.0",
+    "@types/jest": "^26.0.13",
+    "@votingworks/ballot-encoder": "^4.0.0",
+    "faker": "^5.1.0",
+    "i18next": "^19.7.0",
+    "jest": "^26.4.2",
+    "js-sha256": "^0.9.0",
+    "react": "^16.13.1",
+    "react-dom": "^16.13.1",
+    "react-i18next": "^11.6.0",
+    "ts-jest": "^26.3.0",
+    "typescript": "^4.0.2"
+  },
+  "dependencies": {
+    "@votingworks/qrcode.react": "^1.0.1",
+    "base64-js": "^1.3.1"
+  }
+}
