+++ libs/hmpb-ui/src/Prose.tsx
@@ -1,12 +1,12 @@
 import styled from 'styled-components'
 
-interface Props {
+export interface ProseProps {
   compact?: boolean
   maxWidth?: boolean
   textCenter?: boolean
 }
 
-const Prose = styled('div')<Props>`
+export const Prose = styled('div')<ProseProps>`
   margin: ${({ textCenter }) => (textCenter ? 'auto' : undefined)};
 
   /* width: 100%; */
@@ -68,5 +68,3 @@ const Prose = styled('div')<Props>`
     border-top: 0.1em solid #000000;
   }
 `
-
-export default Prose
