+++ libs/hmpb-ui/src/utils/findPartyById.ts
@@ -3,5 +3,3 @@ import { Parties } from '@votingworks/ballot-encoder'
 export const findPartyById = (parties: Parties, id: string) => {
   return parties.find((p) => p.id === id)
 }
-
-export default findPartyById
